Improve interface structure
Changes Made:
In this commit, gradio's interface is re-written in a more modular way, so the project as a whole is better structured, and has a clear entry point (main.py). This makes the project easier to read, work with, maintain, and scale.
