+++ utils/speech.py
